update snippets to new location and syntax
